requirements.txt
pydantic
openai
tiktoken
python-dotenv
